backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
firebase-admin==6.4.0
google-cloud-firestore==2.13.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
spacy==3.7.2
scikit-learn==1.3.2
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.4
numpy==1.24.4
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow==10.1.0
httpx>=0.23.0,<1.0.0
email-validator>=2.0.0
sentence-transformers>=2.2.2
langdetect>=1.0.9